feat(admin): activate an order's balance from the order page, and move the shipment form into a drawer
Two changes to the order detail page.

## 1. A Balance widget, and the admin route behind it

`GET/POST /admin/orders/:id/balance` — read what is owed, and raise it. The
partner portal is where this normally happens, because the makers know when the
goods exist, and dispatch raises it automatically; this is the admin's view of
the same workflow, for an order no partner is working or when someone asks for
it to be done on their behalf.

The widget shows the deposit, the balance and the order total, the buyer's
payment link once one exists, and a single button. It renders NOTHING for an
ordinary order — most orders have no payment schedule, and an empty card on
every order page is noise.

🔴 `POST` requires `confirm: true`. Raising a balance sends a real buyer a link
and creates a charge; it is not undoable by the caller.

🔑 Refusals are shown VERBATIM. "The deposit on this order is pending, so the
balance cannot be raised yet" is the only place an operator learns why the
button is disabled — replacing it with "something went wrong" would throw away
the one useful thing the backend said.

`GET` reconciles first: the payment module emits no events, so an admin opening
the order is another reliable moment to notice money that already landed.

## 2. The carrier shipment form moves into a RouteDrawer

`/app/orders/:id/carrier-shipment`. It is a long form — live rates, dimensions,
pickup slot, label, manual AWB — and as a flat section it pushed the rest of the
order detail below the fold whether or not anyone was shipping that day.

The form is lifted verbatim into
`components/forms/order-carrier-shipment/carrier-shipment-form.tsx`, taking an
`orderId` rather than the widget's `DetailWidgetProps` because it now has two
callers. The widget becomes a door: a heading, a sentence and a button.

⚠️ The widget is deliberately dumb — it fetches nothing. Duplicating any of the
drawer's state would be a second opinion about what has already been booked; the
read-only tracking widget below it is what shows the result.

Nesting an extension route under the CORE order page is the same pattern
`routes/products/[id]/link-people` already uses, so the order detail stays
behind the drawer and the form keeps its own URL — a half-filled booking now
survives a refresh and can be linked to.

Verify: `pnpm check:prod-build` → "Frontend build completed successfully". The
only backend error is the pre-existing `@jest/core` one, which CI tolerates.
